refactor: 궁 안에서의 움직임을 확인하는 PalacePiece 추상클래스 추가
